Require/patch gems in a recipe context
